--- cypress/pageObjects/JobConfigurePage.js
@@ -1,11 +1,0 @@
-import MultiConfigurationProjectConfigurePage from "./MultiConfigurationProjectConfigurePage";
-
-class JobConfigurePage {
-    getSaveButton = () => cy.get("button[name='Submit']");
-
-    clickSaveButton() {
-        this.getSaveButton().click();
-        return new MultiConfigurationProjectConfigurePage();
-    }
-}
-export default JobConfigurePage;
